Added types and swapped names of ClickToPayHolder.tsx & ClickToPayWrapper.tsx
