+++ packages/lib/src/components/Card/ClickToPayWrapper.tsx
@@ -1,67 +1,19 @@
-import { Fragment, h } from 'preact';
-import useClickToPayContext from './components/ClickToPay/context/useClickToPayContext';
-import { CtpState } from './components/ClickToPay/services/ClickToPayService';
-import { useCallback, useEffect, useState } from 'preact/hooks';
-import ClickToPayComponent from './components/ClickToPay';
-import ContentSeparator from '../internal/ContentSeparator';
-import Button from '../internal/Button';
-import useCoreContext from '../../core/Context/useCoreContext';
+import ClickToPayProvider, { ClickToPayProviderProps } from './components/ClickToPay/context/ClickToPayProvider';
+import ClickToPayHolder from './ClickToPayHolder';
+import { h } from 'preact';
 
-type ClickToPayWrapperProps = {
-    children(isCardPrimaryInput?: boolean): h.JSX.Element;
-};
-
-const ClickToPayWrapper = ({ children }: ClickToPayWrapperProps) => {
-    const { i18n } = useCoreContext();
-    const [isCardInputVisible, setIsCardInputVisible] = useState<boolean>(null);
-    const { ctpState, isCtpPrimaryPaymentMethod, setIsCtpPrimaryPaymentMethod, status } = useClickToPayContext();
-
-    const areFieldsNotSet = isCardInputVisible === null && isCtpPrimaryPaymentMethod === null;
-
-    useEffect(() => {
-        if (areFieldsNotSet) {
-            if (ctpState === CtpState.ShopperIdentified || ctpState === CtpState.Ready) {
-                setIsCardInputVisible(false);
-                setIsCtpPrimaryPaymentMethod(true);
-                return;
-            }
-            if (ctpState === CtpState.NotAvailable) {
-                setIsCardInputVisible(true);
-                setIsCtpPrimaryPaymentMethod(false);
-            }
-        }
-    }, [ctpState, areFieldsNotSet]);
-
-    const handleOnShowCardButtonClick = useCallback(() => {
-        setIsCardInputVisible(true);
-        setIsCtpPrimaryPaymentMethod(false);
-    }, []);
-
-    if (ctpState === CtpState.NotAvailable) {
-        return children();
-    }
-
-    if (ctpState === CtpState.Loading || ctpState === CtpState.ShopperIdentified) {
-        return <ClickToPayComponent />;
-    }
-
+const ClickToPayWrapper = ({ amount, clickToPayService, setClickToPayRef, onSetStatus, onSubmit, onError, ...props }: ClickToPayProviderProps) => {
     return (
-        <Fragment>
-            <ClickToPayComponent onShowCardButtonClick={handleOnShowCardButtonClick} />
-
-            <ContentSeparator classNames={['adyen-checkout-ctp__separator']} label={i18n.get('ctp.separatorText')} />
-
-            {isCardInputVisible ? (
-                children(!isCtpPrimaryPaymentMethod)
-            ) : (
-                <Button
-                    variant="secondary"
-                    disabled={status === 'loading'}
-                    label={i18n.get('ctp.manualCardEntry')}
-                    onClick={handleOnShowCardButtonClick}
-                />
-            )}
-        </Fragment>
+        <ClickToPayProvider
+            amount={amount}
+            clickToPayService={clickToPayService}
+            setClickToPayRef={setClickToPayRef}
+            onSetStatus={onSetStatus}
+            onSubmit={onSubmit}
+            onError={onError}
+        >
+            <ClickToPayHolder>{props.children}</ClickToPayHolder>
+        </ClickToPayProvider>
     );
 };
